refactor: move db to separate package
