+++ packages/db/.env.example
@@ -1,0 +1,3 @@
+DATABASE_URL=
+DATABASE_AUTH_TOKEN=
+
